+++ CHANGES.rst
@@ -12,7 +12,13 @@ Changes
 
 Changes:
 --------
-- No change.
+- Improve conformance for returned status codes and error messages when requesting results for an unfinished,
+  failed, or dismissed ``Job``.
+- Adjust conformance item references to correspond with `OGC API - Processes: Part 2` renamed from `Transactions` to
+  `Deploy, Replace, Undeploy`.
+- Add ``mutable`` field to ``Process`` summary listing and detailed descriptions for conformance
+  (resolves `#180 <https://github.com/crim-ca/weaver/issues/180>`_).
+- Improve ``Process`` undeployment to consider running ``Job`` to block its removal while in use.
 
 Fixes:
 ------